requirements.txt
aiohttp==3.11.11
discord.py==2.4.0
python-dotenv==1.0.1
Requests==2.32.3
